[front/assistant_builder] - refactor: rename PromptDiff functionalities to InstructionDiff
 - Replaced PromptDiffExtension with the new InstructionDiffExtension which aligns better with the terminology used elsewhere
 - Updated InstructionScreen component to use InstructionDiffExtension instead of the removed PromptDiffExtension
 - Renamed PromptHistory component to InstructionHistory to maintain consistency with the new naming convention
 - Removed outdated PromptDiffExtension.ts file and its references in the codebase
